requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
arxiv==1.4.8
astroid==3.0.0
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
dataclasses-json==0.6.1
decorator==5.1.1
dill==0.3.7
feedparser==6.0.10
frozenlist==1.4.0
fsspec==2023.9.2
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
isort==5.12.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.311
langsmith==0.0.43
llama-index==0.8.42
magic-filter==1.0.12
marshmallow==3.20.1
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.26.0
openai==0.28.1
packaging==23.2
pandas==2.1.1
pdfminer.six==20221105
platformdirs==3.11.0
py==1.11.0
pyarrow==13.0.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pylint==3.0.1
python-dateutil==2.8.2
python-telegram-bot==20.6
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
retry==0.9.2
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
tenacity==8.2.3
tiktoken==0.5.1
tomlkit==0.12.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.17
yarl==1.9.2
